detector/requirements.txt
pyyaml==6.0.1
requests==2.31.0
psutil==5.9.8
